CICD construct; use in LB service; project structure refactor
