+++ cdk/examples/components/core_infrastructure_construct.py
